requirements.txt
python-telegram-bot==20.7
tinkoff-investments==0.2.0b114
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
tabulate==0.9.0
python-dotenv==1.0.1
redis==5.0.3
sentry-sdk==1.45.0
aiosqlite==0.20.0